# advanced/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CineTrack Advanced API",
    "description": "CineTrack advanced-model API. The Flutter client talks only to this backend. Movie/TV discovery, details, cast/crew, seasons and episodes are fetched from TMDB, normalized and cached server-side. The API also provides PostgreSQL-backed users, JWT sessions, watch state/progress, ratings, spoiler-aware comments, favorites, personal lists, statistics, reports and role-based administration.",
    "contact": {
      "name": "CineTrack Project Team"
    },
    "license": {
      "name": "Educational non-commercial project"
    },
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/me/library": {
      "get": {
        "tags": [
          "Personal library"
        ],
        "summary": "Library",
        "operationId": "library_api_v1_me_library_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "watch_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(plan_to_watch|watching|completed|paused|dropped|favorite)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Watch Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_LibraryEntry_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/library/{media_id}": {
      "put": {
        "tags": [
          "Personal library"
        ],
        "summary": "Update Library",
        "operationId": "update_library_api_v1_me_library__media_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMediaState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Personal library"
        ],
        "summary": "Delete Library",
        "operationId": "delete_library_api_v1_me_library__media_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/favorites/{media_id}": {
      "put": {
        "tags": [
          "Personal library"
        ],
        "summary": "Favorite",
        "operationId": "favorite_api_v1_me_favorites__media_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Personal library"
        ],
        "summary": "Unfavorite",
        "operationId": "unfavorite_api_v1_me_favorites__media_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/ratings/{media_id}": {
      "put": {
        "tags": [
          "Personal library"
        ],
        "summary": "Rate",
        "operationId": "rate_api_v1_me_ratings__media_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RatingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingDistribution"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Personal library"
        ],
        "summary": "Unrate",
        "operationId": "unrate_api_v1_me_ratings__media_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/episodes/{episode_id}": {
      "put": {
        "tags": [
          "Personal library"
        ],
        "summary": "Mark Episode",
        "operationId": "mark_episode_api_v1_me_episodes__episode_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Episode Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpisodeWatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/series/{media_id}/progress": {
      "get": {
        "tags": [
          "Personal library"
        ],
        "summary": "Progress",
        "operationId": "progress_api_v1_me_series__media_id__progress_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EpisodeWatchRequest": {
        "properties": {
          "watched": {
            "type": "boolean",
            "title": "Watched",
            "default": true
          }
        },
        "type": "object",
        "title": "EpisodeWatchRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LibraryEntry": {
        "properties": {
          "media": {
            "$ref": "#/components/schemas/MediaSummary"
          },
          "watch_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Watch Status"
          },
          "is_favorite": {
            "type": "boolean",
            "title": "Is Favorite"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProgressResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "media",
          "watch_status",
          "is_favorite",
          "progress",
          "updated_at"
        ],
        "title": "LibraryEntry"
      },
...
      "MediaSummary": {
        "properties": {
          "media_id": {
            "type": "string",
            "title": "Media Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "tmdb"
          },
          "tmdb_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tmdb Id"
          },
          "imdb_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imdb Id"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "original_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Title"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          },
          "plot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Genres"
          },
          "release_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Year"
          },
          "end_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Year"
          },
          "provider_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Rating"
          },
          "app_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Rating"
          },
          "rating_count": {
            "type": "integer",
            "title": "Rating Count",
            "default": 0
          },
          "release_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Status"
          },
          "season_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season Count"
          },
          "episode_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode Count"
          }
        },
        "type": "object",
        "required": [
          "media_id",
          "media_type",
          "title"
        ],
        "title": "MediaSummary"
      },
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
...
      "PaginatedResponse_LibraryEntry_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LibraryEntry"
            },
            "type": "array",
            "title": "Items"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "items",
          "pagination"
        ],
        "title": "PaginatedResponse[LibraryEntry]"
      },
...
      "PaginationMeta": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "total_pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Pages"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_previous": {
            "type": "boolean",
            "title": "Has Previous"
          }
        },
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total",
          "total_pages",
          "has_next",
          "has_previous"
        ],
        "title": "PaginationMeta"
      },
...
      "ProgressResponse": {
        "properties": {
          "media_id": {
            "type": "string",
            "title": "Media Id"
          },
          "watched_episodes": {
            "type": "integer",
            "title": "Watched Episodes"
          },
          "total_episodes": {
            "type": "integer",
            "title": "Total Episodes"
          },
          "remaining_episodes": {
            "type": "integer",
            "title": "Remaining Episodes"
          },
          "progress_percent": {
            "type": "number",
            "title": "Progress Percent"
          },
          "progress_color": {
            "type": "string",
            "title": "Progress Color"
          },
          "progress_hex": {
            "type": "string",
            "title": "Progress Hex"
          },
          "release_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Status"
          }
        },
        "type": "object",
        "required": [
          "media_id",
          "watched_episodes",
          "total_episodes",
          "remaining_episodes",
          "progress_percent",
          "progress_color",
          "progress_hex",
          "release_status"
        ],
        "title": "ProgressResponse"
      },
      "RatingDistribution": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "percentages": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Percentages"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts"
          }
        },
        "type": "object",
        "required": [
          "total",
          "percentages",
          "counts"
        ],
        "title": "RatingDistribution"
      },
      "RatingRequest": {
        "properties": {
          "score": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "score"
        ],
        "title": "RatingRequest"
      },
...
      "UserMediaState": {
        "properties": {
          "watch_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Watch Status"
          },
          "is_favorite": {
            "type": "boolean",
            "title": "Is Favorite",
            "default": false
          },
          "user_rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Rating"
          },
          "watched_episodes": {
            "type": "integer",
            "title": "Watched Episodes",
            "default": 0
          },
          "remaining_episodes": {
            "type": "integer",
            "title": "Remaining Episodes",
            "default": 0
          },
          "progress_percent": {
            "type": "number",
            "title": "Progress Percent",
            "default": 0
          },
          "progress_color": {
            "type": "string",
            "title": "Progress Color",
            "default": "black"
          },
          "progress_hex": {
            "type": "string",
            "title": "Progress Hex",
            "default": "#212121"
          }
        },
        "type": "object",
        "title": "UserMediaState"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WatchStatusRequest": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "WatchStatusRequest"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}